Refatoração BigDecimal
De Double para BigDecimal

+++ src/java/beans/ItemVenda.java
@@ -1,6 +1,7 @@
 package beans;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import javax.persistence.Entity;
 import javax.persistence.Id;
@@ -12,13 +13,13 @@ public class ItemVenda implements Serializable {
     private Long id;
     private Long idProduto;
     private Long idVenda;
-    private Double precoCompra;
-    private Double precoVenda;
-    private Double quantidade;
-    private Double porcentagemDesconto;
+    private BigDecimal precoCompra;
+    private BigDecimal precoVenda;
+    private BigDecimal quantidade;
+    private BigDecimal porcentagemDesconto;
     private Boolean sugestao;
     private Timestamp dataGarantia;
-    private Double porcentagemGarantia;
+    private BigDecimal porcentagemGarantia;
 
     public Long getId() {
         return id;
@@ -44,35 +45,35 @@ public class ItemVenda implements Serializable {
         this.idVenda = idVenda;
     }
 
-    public Double getPrecoCompra() {
+    public BigDecimal getPrecoCompra() {
         return precoCompra;
     }
 
-    public void setPrecoCompra(Double precoCompra) {
+    public void setPrecoCompra(BigDecimal precoCompra) {
         this.precoCompra = precoCompra;
     }
 
-    public Double getPrecoVenda() {
+    public BigDecimal getPrecoVenda() {
         return precoVenda;
     }
 
-    public void setPrecoVenda(Double precoVenda) {
+    public void setPrecoVenda(BigDecimal precoVenda) {
         this.precoVenda = precoVenda;
     }
 
-    public Double getQuantidade() {
+    public BigDecimal getQuantidade() {
         return quantidade;
     }
 
-    public void setQuantidade(Double quantidade) {
+    public void setQuantidade(BigDecimal quantidade) {
         this.quantidade = quantidade;
     }
 
-    public Double getPorcentagemDesconto() {
+    public BigDecimal getPorcentagemDesconto() {
         return porcentagemDesconto;
     }
 
-    public void setPorcentagemDesconto(Double porcentagemDesconto) {
+    public void setPorcentagemDesconto(BigDecimal porcentagemDesconto) {
         this.porcentagemDesconto = porcentagemDesconto;
     }
 
@@ -92,11 +93,11 @@ public class ItemVenda implements Serializable {
         this.dataGarantia = dataGarantia;
     }
 
-    public Double getPorcentagemGarantia() {
+    public BigDecimal getPorcentagemGarantia() {
         return porcentagemGarantia;
     }
 
-    public void setPorcentagemGarantia(Double porcentagemGarantia) {
+    public void setPorcentagemGarantia(BigDecimal porcentagemGarantia) {
         this.porcentagemGarantia = porcentagemGarantia;
     }
